Unify data and template of field resolvers and operations
Conceptually, GraphQL operations (queries/mutations/subscriptions) are a
special case of field resolvers that happen to be on well known root types.

Technically, this means that everything about field resolvers is the same
as these root types resolvers, apart from the fact that the former get an
instance of the resolved type as first param.

Therefore, all the configuration that applies to field resolvers also
applies to resolvers of the fields that are part of these root types, like
CompletableFuture return type, or DataFetcherEnvironment parameter.

Resolves: https://github.com/kobylynskyi/graphql-java-codegen/issues/64
Resolves: https://github.com/kobylynskyi/graphql-java-codegen/issues/61
Resolves: https://github.com/kobylynskyi/graphql-java-codegen/issues/63