Add annotations to JavaBean getters
Add nonnull and other annotations to the getter method in addition to the field declaration. Some Java Bean introspection tools can only use annotations declared on the getter/setter methods, not on the private fields, so we need to repeat the annotation that was placed on the field on the getter as well to get full introspection with annotations.